fix: detection of process liveness on Windows
