+++ pkg/oci/builder.go
@@ -6,7 +6,6 @@ import (
 	"os"
 	"path/filepath"
 	"strconv"
-	"syscall"
 	"time"
 
 	v1 "github.com/google/go-containerregistry/pkg/v1"
@@ -224,19 +223,6 @@ func teardown(cfg *buildConfig) {
 		}
 		_ = os.RemoveAll(dir)
 	}
-}
-
-func processExists(pid string) bool {
-	p, err := strconv.Atoi(pid)
-	if err != nil {
-		return false
-	}
-	process, err := os.FindProcess(p)
-	if err != nil {
-		return false
-	}
-	err = process.Signal(syscall.Signal(0))
-	return err == nil
 }
 
 func isLinkTo(link, target string) bool {
