+++ pkg/oci/process_exists_unix.go
@@ -1,0 +1,22 @@
+//go:build linux || darwin
+
+package oci
+
+import (
+	"os"
+	"strconv"
+	"syscall"
+)
+
+func processExists(pid string) bool {
+	p, err := strconv.Atoi(pid)
+	if err != nil {
+		return false
+	}
+	process, err := os.FindProcess(p)
+	if err != nil {
+		return false
+	}
+	err = process.Signal(syscall.Signal(0))
+	return err == nil
+}
