+++ pkg/oci/process_exists_windows.go
@@ -1,0 +1,18 @@
+package oci
+
+import (
+	"os"
+	"strconv"
+)
+
+func processExists(pid string) bool {
+	p, err := strconv.Atoi(pid)
+	if err != nil {
+		return false
+	}
+	_, err = os.FindProcess(p)
+	if err != nil {
+		return false
+	}
+	return true
+}
